Add Daemon Class AsyncTask is great at handling single tasks in the background, but some times you will want to continue having a small background process running. For this, a Thread is more usefull, but they should be a lot easier to add while having the most basic operations.
The Daemon task is not really a daemon, as this will pause and resume whenever your Activity does so. It also provides tools to do a complete stop and restart and it automatically handles the loop. You just create the Run method in the abstract class or pass a Runnable, then execute the start method while parsing the timeout in milliseconds